requirements.txt
google.generativeai
numpy
flask
pandas
sentence-transformers